Refactor background jobs pattern for docs
